Begin multiprocessing misadventure
There is no per-instance state involved in USE_SHELL, so pickling
is far less directly relevant than usual to multiprocessing: the
spawn and forkserver methods will not preserve a subsequently
changed attribute value unless side effects of loading a module (or
other unpickling of a function or its arguments that are submitted
to run on a worker subprocess) causes it to run again; the fork
method will.

This will be (automatically) the same with any combination of
metaclasses, properties, and custom descriptors as in the more
straightforward case of a simple class attribute. Subtleties arise
in the code that uses GitPython and multiprocessing, but should not
arise unintentionally from the change in implementation of USE_SHELL
done to add deprecation warnings, except possibly with respect to
whether warnings will be repeated in worker processes, which is
less important than whether the actual state is preserved.